requirements.txt
Pillow
torch
torchvision
numpy
scikit-learn
httpx
pymongo
tritonclient
gunicorn
flask